refactor: split memory integration tests by behavior
